Move sessionAttributes folder
