+++ backend-api/src/functions/common/session/getOperations/TxmaEvent/TxMAEvent.ts
@@ -2,7 +2,7 @@ import { AttributeValue } from "@aws-sdk/client-dynamodb";
 import { marshall } from "@aws-sdk/util-dynamodb";
 import { Result } from "../../../../utils/result";
 import { SessionAttributes } from "../../session";
-import { getBiometricTokenIssuedSessionAttributes } from "../../updateOperations/sessionAttributes/sessionAttributes";
+import { getBiometricTokenIssuedSessionAttributes } from "../../sessionAttributes/sessionAttributes";
 import { GetSessionOperation } from "../GetSessionOperation";
 
 export class TxMAEvent implements GetSessionOperation {
